+++ src/features/customColumns/CustomColumnsTypes.ts
@@ -1,0 +1,4 @@
+export type ColumnT = {
+    name: string;
+    type: 'general' | 'metadata';
+};